+++ src/main/java/com/klin/holoItems/Events.java
@@ -3,9 +3,9 @@ package com.klin.holoItems;
 import com.klin.holoItems.abstractClasses.Enchant;
 import com.klin.holoItems.abstractClasses.Pack;
 import com.klin.holoItems.abstractClasses.Wiring;
-import com.klin.holoItems.collections.gamers.okayuCollection.AutoEat;
-import com.klin.holoItems.collections.gamers.okayuCollection.BadAutoEat;
-import com.klin.holoItems.collections.gamers.okayuCollection.BasicAutoEat;
+import com.klin.holoItems.collections.gamers.okayuCollection.items.AutoEat;
+import com.klin.holoItems.collections.gamers.okayuCollection.items.BadAutoEat;
+import com.klin.holoItems.collections.gamers.okayuCollection.items.BasicAutoEat;
 import com.klin.holoItems.collections.gen1.melCollection.items.ReadingGlasses;
 import com.klin.holoItems.collections.gen2.shionCollection.items.Fireball;
 import com.klin.holoItems.collections.gen2.shionCollection.items.SecretBrew;
