+++ src/main/java/com/klin/holoItems/collections/gamers/okayuCollection/items/AutoEat.java
@@ -1,12 +1,12 @@
-package com.klin.holoItems.collections.gamers.okayuCollection;
+package com.klin.holoItems.collections.gamers.okayuCollection.items;
 
 import com.klin.holoItems.Collections;
 import com.klin.holoItems.HoloItems;
 import com.klin.holoItems.abstractClasses.Enchant;
+import com.klin.holoItems.collections.gamers.okayuCollection.OkayuCollection;
 import com.klin.holoItems.collections.gen1.melCollection.items.ReadingGlasses;
 import com.klin.holoItems.collections.gen4.cocoCollection.items.DragonHorns;
 import com.klin.holoItems.collections.gen5.botanCollection.items.Backdash;
-import com.klin.holoItems.collections.misc.ingredientsCollection.items.CoalPetal;
 import com.klin.holoItems.interfaces.Hungerable;
 import com.klin.holoItems.utility.Utility;
 import it.unimi.dsi.fastutil.objects.ObjectObjectMutablePair;
@@ -16,7 +16,6 @@ import org.bukkit.NamespacedKey;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
 import org.bukkit.inventory.*;
 
-import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
