+++ src/main/java/com/klin/holoItems/collections/gamers/okayuCollection/items/BasicAutoEat.java
@@ -1,12 +1,12 @@
-package com.klin.holoItems.collections.gamers.okayuCollection;
+package com.klin.holoItems.collections.gamers.okayuCollection.items;
 
 import com.klin.holoItems.Collections;
 import com.klin.holoItems.HoloItems;
 import com.klin.holoItems.abstractClasses.Enchant;
+import com.klin.holoItems.collections.gamers.okayuCollection.OkayuCollection;
 import com.klin.holoItems.collections.gen1.melCollection.items.ReadingGlasses;
 import com.klin.holoItems.collections.gen4.cocoCollection.items.DragonHorns;
 import com.klin.holoItems.collections.gen5.botanCollection.items.Backdash;
-import com.klin.holoItems.collections.misc.ingredientsCollection.items.CoalPetal;
 import com.klin.holoItems.interfaces.Hungerable;
 import com.klin.holoItems.utility.Utility;
 import org.bukkit.Bukkit;
