+++ src/main/java/com/klin/holoItems/collections/gamers/okayuCollection/items/GoodAutoEat.java
@@ -1,16 +1,14 @@
-package com.klin.holoItems.collections.gamers.okayuCollection;
+package com.klin.holoItems.collections.gamers.okayuCollection.items;
 
 import com.klin.holoItems.Collections;
 import com.klin.holoItems.HoloItems;
 import com.klin.holoItems.abstractClasses.Enchant;
+import com.klin.holoItems.collections.gamers.okayuCollection.OkayuCollection;
 import com.klin.holoItems.collections.gen1.melCollection.items.ReadingGlasses;
 import com.klin.holoItems.collections.gen4.cocoCollection.items.DragonHorns;
 import com.klin.holoItems.collections.gen5.botanCollection.items.Backdash;
-import com.klin.holoItems.collections.misc.ingredientsCollection.items.CoalPetal;
 import com.klin.holoItems.interfaces.Hungerable;
 import com.klin.holoItems.utility.Utility;
-import it.unimi.dsi.fastutil.objects.ObjectObjectMutablePair;
-import jdk.jshell.execution.Util;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
